Allow a build program to be an external command
This allows custom build commands to be specified for files without
needing to extend rstblog for every external command that you might want
to run.  External commands start with an exclamation mark and must have
two arguments which begin with a percent sign - these are the input and
output file names in order.

Filename arguments are allowed to have a suffix so that file extensions
can be changed, for example by adding the follwing to a "config.yml"
file in a directory, files in that directory are compiled by "lessc" and
".less" files are renamed to ".css":

    program: "!lessc %.less %.css"